chore: Move constants
