+++ mobile/src/components/Form/Button/Icon.tsx
@@ -4,7 +4,7 @@ import { View } from "react-native";
 import type { Icon } from "~/resources/icons/type";
 
 import { cn } from "~/lib/style";
-import type { AppColor } from "~/modules/customization/theme/constants";
+import type { AppColor } from "~/modules/customization/theme/core/constants";
 import { useColor } from "~/modules/customization/theme/hooks";
 import type { PressProps } from "./types";
 import { Pressable } from "../../Base/Pressable";
