+++ mobile/src/lib/style.ts
@@ -4,7 +4,7 @@ import { extendTailwindMerge } from "tailwind-merge";
 
 import { BorderRadius, FontSize } from "~/constants/Styles";
 import { FontFamily } from "~/modules/customization/font/core/constants";
-import { ColorRoleOptions } from "~/modules/customization/theme/constants";
+import { ColorRoleOptions } from "~/modules/customization/theme/core/constants";
 
 //#region Color
 // Need to include `transparent` as otherwise, things will get merged incorrectly.
