+++ mobile/src/modules/customization/theme/core/constants.ts
