+++ mobile/src/modules/media/components/ArtistsLink.tsx
@@ -10,7 +10,7 @@ import { cn } from "~/lib/style";
 import { Pressable } from "~/components/Base/Pressable";
 import { Marquee } from "~/components/Marquee";
 import { StyledText } from "~/components/Typography/StyledText";
-import type { ColorRole } from "~/modules/customization/theme/constants";
+import type { ColorRole } from "~/modules/customization/theme/core/constants";
 import type { PopStrategy } from "~/stores/Session/types";
 
 /** Renders display string for artists, with different onPress actions based on the number of artists. */
