+++ mobile/src/modules/search/components/SearchList.tsx
@@ -10,7 +10,7 @@ import { cn } from "~/lib/style";
 import type { FlatListProps, ListRenderItemInfo } from "~/components/Base/List";
 import { FlatList } from "~/components/Base/List";
 import { TopDownGradient } from "~/components/Gradient";
-import type { ColorRole } from "~/modules/customization/theme/constants";
+import type { ColorRole } from "~/modules/customization/theme/core/constants";
 import { SearchBar } from "./SearchBar";
 
 const DEFAULT_GAP = 24;
