+++ mobile/src/stores/Preference/constants.ts
@@ -2,7 +2,7 @@ import type { Font } from "~/modules/customization/font/core/constants";
 import type {
   CustomTheme,
   DefaultTheme,
-} from "~/modules/customization/theme/constants";
+} from "~/modules/customization/theme/core/constants";
 import type { Tab } from "./types";
 
 //#region Now Playing Design
